+++ src/test/java/org/folio/support/base/TestConstants.java
@@ -27,8 +27,6 @@ public class TestConstants {
   public static final String SOURCE_FILE_CODE = "sourceFileType";
   public static final String SOURCE_FILE_NATURAL_ID = "sourceFileNaturalId";
   public static final String SOURCE_FILE_SOURCE = "sourceFileSource";
-  public static final String AUTHORITY_SOURCE = "SOURCE";
-  public static final String AUTHORITY_CONSORTIUM_SOURCE = CONSORTIUM_SOURCE_PREFIX + AUTHORITY_SOURCE;
   public static final String TEST_PROPERTY_VALUE = "value";
   public static final UUID TEST_ID = UUID.randomUUID();
   public static final int TEST_VERSION = 2;
@@ -68,6 +66,7 @@ public class TestConstants {
   private static final String AUTH_STATS_ENDPOINT_PATH_PATTERN = "/links/stats/authority";
   private static final String AUTH_STATS_ENDPOINT_PARAMS = "?action=%s&fromDate=%s&toDate=%s&limit=%d";
   private static final String LINKING_RULES_ENDPOINT = "/linking-rules/instance-authority";
+  private static final String AUTHORITY_HEADING_TYPES_ENDPOINT = "/authority-heading-types";
   private static final String AUTHORITY_IDENTIFIER_TYPES_ENDPOINT = "/authority-identifier-types";
   private static final String AUTHORITY_NOTE_TYPES_ENDPOINT = "/authority-note-types";
   private static final String AUTHORITY_SOURCE_FILES_ENDPOINT = "/authority-source-files";
@@ -165,6 +164,10 @@ public class TestConstants {
 
   public static String authorityNoteTypesEndpoint(UUID id) {
     return AUTHORITY_NOTE_TYPES_ENDPOINT + "/" + id;
+  }
+
+  public static String authorityHeadingTypesEndpoint() {
+    return AUTHORITY_HEADING_TYPES_ENDPOINT;
   }
 
   public static String authorityIdentifierTypesEndpoint() {